split data_finder.helpers into smaller chunks
no functional changes in this commit
just moving stuff about
