+++ polling_stations/apps/data_finder/tests/test_addressbase_geocoder_adapter.py
@@ -1,6 +1,6 @@
 from django.core.exceptions import ObjectDoesNotExist
 from django.test import TestCase
-from data_finder.helpers import (
+from data_finder.helpers.geocoders import (
     AddressBaseGeocoderAdapter, MultipleCouncilsException)
 from uk_geo_utils.geocoders import (
     AddressBaseGeocoder,
