+++ polling_stations/apps/data_finder/tests/test_geocoders.py
@@ -1,6 +1,6 @@
 import mock
 from django.test import TestCase
-from data_finder.helpers import (
+from data_finder.helpers.geocoders import (
     geocode, geocode_point_only, OnspdGeocoderAdapter, MultipleCouncilsException
 )
 from uk_geo_utils.geocoders import AddressBaseGeocoder, OnspdGeocoder
@@ -25,7 +25,7 @@ class GeocodeTest(TestCase):
 
     fixtures = ['test_addressbase.json']
 
-    @mock.patch("data_finder.helpers.OnspdGeocoderAdapter.geocode", mock_geocode)
+    @mock.patch("data_finder.helpers.geocoders.OnspdGeocoderAdapter.geocode", mock_geocode)
     def test_no_records(self):
         """
         We can't find any records for the given postcode in the AddressBase table
@@ -35,7 +35,7 @@ class GeocodeTest(TestCase):
         result = geocode('DD1 1DD')
         self.assertIsInstance(result, OnspdGeocoder)
 
-    @mock.patch("data_finder.helpers.OnspdGeocoderAdapter.geocode", mock_geocode)
+    @mock.patch("data_finder.helpers.geocoders.OnspdGeocoderAdapter.geocode", mock_geocode)
     def test_no_codes(self):
         """
         We find records for the given postcode in the AddressBase table
@@ -46,7 +46,7 @@ class GeocodeTest(TestCase):
         result = geocode('AA11AA')
         self.assertIsInstance(result, OnspdGeocoder)
 
-    @mock.patch("data_finder.helpers.OnspdGeocoderAdapter.geocode", mock_geocode)
+    @mock.patch("data_finder.helpers.geocoders.OnspdGeocoderAdapter.geocode", mock_geocode)
     def test_multiple_councils(self):
         """
         We find records for the given postcode in the AddressBase table
@@ -62,7 +62,7 @@ class GeocodeTest(TestCase):
             exception_thrown = True
         self.assertTrue(exception_thrown)
 
-    @mock.patch("data_finder.helpers.OnspdGeocoderAdapter.geocode", mock_geocode)
+    @mock.patch("data_finder.helpers.geocoders.OnspdGeocoderAdapter.geocode", mock_geocode)
     def test_valid(self):
         """
         We find records for the given postcode in the AddressBase table
@@ -78,7 +78,7 @@ class GeocodePointOnlyTest(TestCase):
 
     fixtures = ['test_addressbase.json']
 
-    @mock.patch("data_finder.helpers.OnspdGeocoderAdapter.geocode_point_only", mock_geocode)
+    @mock.patch("data_finder.helpers.geocoders.OnspdGeocoderAdapter.geocode_point_only", mock_geocode)
     def test_no_records(self):
         """
         We can't find any records for the given postcode in the AddressBase table
@@ -88,7 +88,7 @@ class GeocodePointOnlyTest(TestCase):
         result = geocode_point_only('DD1 1DD')
         self.assertIsInstance(result, OnspdGeocoder)
 
-    @mock.patch("data_finder.helpers.OnspdGeocoderAdapter.geocode_point_only", mock_geocode)
+    @mock.patch("data_finder.helpers.geocoders.OnspdGeocoderAdapter.geocode_point_only", mock_geocode)
     def test_valid(self):
         """
         We find records for the given postcode in the AddressBase table
