+++ polling_stations/apps/data_finder/tests/test_google_directions.py
@@ -2,7 +2,7 @@ import json
 from django.conf import settings
 from django.contrib.gis.geos import Point
 from django.test import TestCase
-from data_finder.directions_clients import (
+from data_finder.helpers.directions import (
     DirectionsException, GoogleDirectionsClient)
